move: 型定義ファイルをtypes/signup/に移動しリネームした
